chore(introspection): move pg specific tests to postgres modules
