+++ introspection-engine/introspection-engine-tests/tests/enums/mod.rs
@@ -1,9 +1,10 @@
 mod cockroachdb;
+mod mysql;
+mod postgres;
 
 use barrel::types;
-use indoc::indoc;
 use introspection_engine_tests::{test_api::*, TestResult};
-use quaint::prelude::{Queryable, SqlFamily};
+use quaint::prelude::Queryable;
 use test_macros::test_connector;
 
 #[test_connector(exclude(CockroachDb), capabilities(Enums))]
@@ -11,13 +12,11 @@ async fn a_table_with_enums(api: &TestApi) -> TestResult {
     let sql_family = api.sql_family();
 
     if sql_family.is_postgres() {
-        api.database()
-            .raw_cmd(r#"CREATE TYPE "color" AS ENUM ('black', 'white')"#)
-            .await?;
-
-        api.database()
-            .raw_cmd(r#"CREATE TYPE "color2" AS ENUM ('black2', 'white2')"#)
-            .await?;
+        let sql = r#"
+            CREATE TYPE "color" AS ENUM ('black', 'white');
+            CREATE TYPE "color2" AS ENUM ('black2', 'white2');
+        "#;
+        api.database().raw_cmd(sql).await?;
     }
 
     api.barrel()
@@ -109,6 +108,59 @@ async fn a_table_with_an_enum_default_value_that_is_an_empty_string(api: &TestAp
         }}
     "#,
         color,
+    );
+
+    api.assert_eq_datamodels(&dm, &api.introspect().await?);
+
+    Ok(())
+}
+
+
+#[test_connector(exclude(CockroachDb), capabilities(Enums))]
+async fn a_table_with_enum_default_values_that_look_like_booleans(api: &TestApi) -> TestResult {
+    let sql_family = api.sql_family();
+
+    if sql_family.is_postgres() {
+        api.database()
+            .raw_cmd("CREATE Type truth as ENUM ('true', 'false', 'rumor')")
+            .await?;
+    }
+
+    api.barrel()
+        .execute(move |migration| {
+            migration.create_table("News", move |t| {
+                t.add_column("id", types::primary());
+
+                let typ = if sql_family.is_postgres() {
+                    "truth"
+                } else {
+                    "ENUM ('true', 'false', 'rumor')"
+                };
+
+                t.add_column("confirmed", types::custom(typ).nullable(false).default("true"));
+            });
+        })
+        .await?;
+
+    let enum_name = if sql_family.is_mysql() {
+        "News_confirmed"
+    } else {
+        "truth"
+    };
+
+    let dm = format!(
+        r#"
+        model News {{
+            id          Int @id @default(autoincrement())
+            confirmed   {0} @default(true)
+        }}
+        enum {0} {{
+            true
+            false
+            rumor
+        }}
+    "#,
+        enum_name,
     );
 
     api.assert_eq_datamodels(&dm, &api.introspect().await?);
@@ -222,130 +274,6 @@ async fn a_table_with_enum_default_values(api: &TestApi) -> TestResult {
     );
 
     api.assert_eq_datamodels(&dm, &api.introspect().await?);
-
-    Ok(())
-}
-
-#[test_connector(exclude(CockroachDb), capabilities(Enums, ScalarLists))]
-async fn a_table_enums_array(api: &TestApi) -> TestResult {
-    let sql_family = api.sql_family();
-
-    match sql_family {
-        SqlFamily::Postgres => {
-            api.database()
-                .raw_cmd(r#"CREATE Type "color" as ENUM ('black','white')"#)
-                .await?;
-        }
-        _ => todo!("{}", sql_family),
-    }
-
-    api.barrel()
-        .execute(|migration| {
-            migration.create_table("Book", |t| {
-                t.add_column("id", types::primary());
-                t.add_column("color", types::custom("color[]"));
-            });
-        })
-        .await?;
-
-    let dm = indoc! {
-        r#"
-        model Book {
-            id      Int     @id @default(autoincrement())
-            color   color[]
-        }
-
-        enum color {
-            black
-            white
-        }
-        "#,
-    };
-
-    let result = api.introspect().await?;
-
-    api.assert_eq_datamodels(dm, &result);
-
-    Ok(())
-}
-
-#[test_connector(exclude(CockroachDb), capabilities(Enums))]
-async fn a_table_with_enum_default_values_that_look_like_booleans(api: &TestApi) -> TestResult {
-    let sql_family = api.sql_family();
-
-    if sql_family.is_postgres() {
-        api.database()
-            .raw_cmd("CREATE Type truth as ENUM ('true', 'false', 'rumor')")
-            .await?;
-    }
-
-    api.barrel()
-        .execute(move |migration| {
-            migration.create_table("News", move |t| {
-                t.add_column("id", types::primary());
-
-                let typ = if sql_family.is_postgres() {
-                    "truth"
-                } else {
-                    "ENUM ('true', 'false', 'rumor')"
-                };
-
-                t.add_column("confirmed", types::custom(typ).nullable(false).default("true"));
-            });
-        })
-        .await?;
-
-    let enum_name = if sql_family.is_mysql() {
-        "News_confirmed"
-    } else {
-        "truth"
-    };
-
-    let dm = format!(
-        r#"
-        model News {{
-            id          Int @id @default(autoincrement())
-            confirmed   {0} @default(true)
-        }}
-
-        enum {0} {{
-            true
-            false
-            rumor
-        }}
-    "#,
-        enum_name,
-    );
-
-    api.assert_eq_datamodels(&dm, &api.introspect().await?);
-
-    Ok(())
-}
-
-#[test_connector(tags(Postgres))]
-async fn enum_reintrospection_preserves_good_indentation(api: &TestApi) -> TestResult {
-    let original = indoc!(
-        r#"
-        enum MyEnum {
-          A
-          B
-
-          @@map("theEnumName")
-        }
-        "#
-    );
-
-    api.raw_cmd(r#"CREATE TYPE "theEnumName" AS ENUM ('A', 'B');"#).await;
-
-    let reintrospected: String = api
-        .re_introspect(original)
-        .await?
-        .lines()
-        .skip_while(|l| !l.starts_with("enum"))
-        .collect::<Vec<&str>>()
-        .join("\n");
-
-    assert_eq!(original.trim_end(), reintrospected);
 
     Ok(())
 }
